KavachEye-backend/requirements.txt
flask==2.3.3
werkzeug==2.3.7
flask-cors==3.0.10
opencv-python-headless==4.8.0.76
numpy==1.26.4
python-dotenv==0.19.0
requests==2.26.0
pillow>=9.0.0
PyJWT==2.8.0
supabase==2.0.2 